server/requirements.txt
cffi==1.14.0
click==7.1.2
enum34==1.1.10
Flask==1.1.2
Flask-Cors==3.0.8
Flask-MySQLdb==0.2.0
gunicorn==19.10.0
idna==2.9
ipaddress==1.0.23
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
six==1.15.0
Werkzeug==1.0.1